feat(scripts): improve portability with BSD-specific implementations
Replace GNU coreutils commands (gdate, gshred, gshuf) on BSD systems
with built-in implementations to improve portability.

Instead of requiring the installation of GNU coreutils on BSD systems,
the scripts now use built-in commands wrapped in one-line functions to
address discrepancies in commands and options.

The format for these functions is quite simple. For example:

  <internal_prefix><cmd_name><number>() { <body>; }

The script can then call this internal function whenever necessary.