style(scripts): adopt newline before brace for functions
This helps to maintain consistency across my projects. (Yes, it is a
matter of personal taste and consistency at the end of the day).

GNU Bash Manual, Linux Kernel, and FreeBSD follow this style for shell
script functions. Shell builtin 'type' and 'which' command also use
this style for function definition output.