feat(ttybkp): set new backup name format with ttysha
This adds the truncated SHA256 digest of the file to the backup naming
format. Thus:

  {domain}_{scope}_{date_ISO8601}_{SHA256_digest}.{extension}

I'm keeping underscores as separators because being harder to type is
a feature when dealing with backup files (they demand more active
attention). The alternatives--dashes--can cause problems in edge‑case
scenarios, even though I still sanitize and delimit options (--)
anyway.

It is possible to use 'ttysha -s _ ' to adapt backup files with the
old name format to the new one.